--- keyboards/asmodus/plaquead/rgb/keymaps/default/keymap.c
@@ -1,78 +1,0 @@
-/*
-Copyright 2024 Asmodus
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#include QMK_KEYBOARD_H
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-  [0] = LAYOUT_all(
-    KC_ESC,  KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_F6,   KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,  KC_F13,                         KC_PSCR, KC_SCRL, KC_PAUS,
-    KC_GRV,  KC_1,    KC_2,    KC_3,    KC_4,    KC_5,    KC_6,    KC_7,    KC_8,    KC_9,    KC_0,    KC_MINS, KC_EQL,  KC_BSPC, KC_BSPC,               KC_INS,  KC_HOME, KC_PGUP,
-    KC_TAB,  KC_Q,    KC_W,    KC_E,    KC_R,    KC_T,    KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,    KC_LBRC, KC_RBRC, KC_BSLS,                        KC_DEL,  KC_END,  KC_PGDN,
-    KC_CAPS, KC_A,    KC_S,    KC_D,    KC_F,    KC_G,    KC_H,    KC_J,    KC_K,    KC_L,    KC_SCLN, KC_QUOT,    KC_ENT,
-    KC_LSFT, KC_NUBS, KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,    KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_SLSH, KC_RSFT, KC_TRNS,                                 KC_UP,
-    KC_LCTL, KC_LGUI, KC_LALT,                   KC_SPC,                                      KC_RALT, KC_TRNS, KC_RGUI, KC_RCTL,                        KC_LEFT, KC_DOWN, KC_RGHT)
-};
-
-// Code for indicator
-#if defined(RGB_MATRIX_ENABLE) && defined(CAPS_LOCK_LED_INDEX) && defined(SCROLL_LOCK_LED_INDEX)
-
-#if !defined(CAPS_LOCK_MAX_BRIGHTNESS)
-    #define CAPS_LOCK_MAX_BRIGHTNESS RGB_MATRIX_MAXIMUM_BRIGHTNESS
-#endif
-
-#if !defined(CAPS_LOCK_VAL_STEP)
-    #define CAPS_LOCK_VAL_STEP RGB_MATRIX_VAL_STEP
-#endif
-
-#if !defined(SCROLL_LOCK_MAX_BRIGHTNESS)
-    #define SCROLL_LOCK_MAX_BRIGHTNESS RGB_MATRIX_MAXIMUM_BRIGHTNESS
-#endif
-
-#if !defined(SCROLL_LOCK_VAL_STEP)
-    #define SCROLL_LOCK_VAL_STEP RGB_MATRIX_VAL_STEP
-#endif
-
-
-bool rgb_matrix_indicators_user(void) {
-    if (host_keyboard_led_state().caps_lock) {
-        uint8_t b = rgb_matrix_get_val();
-        if (b < CAPS_LOCK_VAL_STEP) {
-            b = CAPS_LOCK_VAL_STEP;
-        } else if (b < (CAPS_LOCK_MAX_BRIGHTNESS - CAPS_LOCK_VAL_STEP)) {
-            b += CAPS_LOCK_VAL_STEP;  // one step more than current brightness
-        } else {
-            b = CAPS_LOCK_MAX_BRIGHTNESS;
-        }
-        rgb_matrix_set_color(CAPS_LOCK_LED_INDEX, b, b, b);  // white, with the adjusted brightness
-    }
-        if (host_keyboard_led_state().scroll_lock) {
-        uint8_t b = rgb_matrix_get_val();
-        if (b < SCROLL_LOCK_VAL_STEP) {
-            b = SCROLL_LOCK_VAL_STEP;
-        } else if (b < (SCROLL_LOCK_MAX_BRIGHTNESS - SCROLL_LOCK_VAL_STEP)) {
-            b += SCROLL_LOCK_VAL_STEP;  // one step more than current brightness
-        } else {
-            b = SCROLL_LOCK_MAX_BRIGHTNESS;
-        }
-        rgb_matrix_set_color(SCROLL_LOCK_LED_INDEX, b, b, b);  // white, with the adjusted brightness
-    }
-    return false;
-}
-
-#endif
-
-// Code for Scroll Locks indicator
-
-
